+++ lib/generators/hyrax/templates/config/initializers/hyrax.rb
@@ -245,8 +245,14 @@ Hyrax.config do |config|
   # Identify the form that will be used for Admin Sets
   # config.administrative_set_form = Hyrax::Forms::AdministrativeSetForm
 
+  # Identify the indexer that will be used for Admin Sets
+  # config.administrative_set_indexer = Hyrax::Indexers::AdministrativeSetIndexer
+
   # Identify the form that will be used for File Sets
   # config.file_set_form = Hyrax::Forms::FileSetForm
+
+  # Identify the indexer that will be used for File Sets
+  # config.file_set_indexer = Hyrax::Indexers::FileSetIndexer
 
   # Service for resolving {FileMetadata} nodes by status, e.g. "primary", rather
   # than use.
@@ -255,9 +261,17 @@ Hyrax.config do |config|
   # Identify the form that will be used for Collections
   # config.pcdm_collection_form = Hyrax::Forms::PcdmCollectionForm
 
+  # Identify the indexer that will be used for Collections
+  # config.pcdm_collection_indexer = Hyrax::Indexers::PcdmCollectionIndexer
+
   # Provide a proc for form generation for Objects
   # config.pcdm_object_form_builder = lambda do |model_class|
   #   Hyrax::Forms::PcdmObjectForm(model_class)
+  # end
+
+  # Provide a proc for indexer generation for Objects
+  # config.pcdm_object_indexer_builder = lambda do |model_class|
+  #   Hyrax::Indexers::PcdmObjectIndexer(model_class)
   # end
 
   # When your application is ready to use the valkyrie index instead of the one
