+++ spec/indexers/hyrax/indexers/pcdm_collection_indexer_spec.rb
@@ -2,7 +2,7 @@
 
 require 'hyrax/specs/shared_specs'
 
-RSpec.describe Hyrax::PcdmCollectionIndexer do
+RSpec.describe Hyrax::Indexers::PcdmCollectionIndexer do
   let(:resource) { FactoryBot.valkyrie_create(:hyrax_collection, :as_collection_member) }
   let(:indexer_class) { described_class }
